+++ src/engine/evaluate.go
@@ -5,11 +5,11 @@ const (
 	BLACK = 1
 )
 
-var mgValue = [6]int16{82, 337, 365, 477, 1025, 0}
-var egValue = [6]int16{94, 281, 297, 512, 936, 0}
-var gamephaseInc = [6]int16{0, 1, 1, 2, 4, 0}
+var mgValue = [6]int{82, 337, 365, 477, 1025, 0}
+var egValue = [6]int{94, 281, 297, 512, 936, 0}
+var gamephaseInc = [6]int{0, 1, 1, 2, 4, 0}
 
-var mgPawnTable = [64]int16{
+var mgPawnTable = [64]int{
 	0, 0, 0, 0, 0, 0, 0, 0,
 	98, 134, 61, 95, 68, 126, 34, -11,
 	-6, 7, 26, 31, 65, 56, 25, -20,
@@ -19,7 +19,7 @@ var mgPawnTable = [64]int16{
 	-35, -1, -20, -23, -15, 24, 38, -22,
 	0, 0, 0, 0, 0, 0, 0, 0}
 
-var egPawnTable = [64]int16{
+var egPawnTable = [64]int{
 	0, 0, 0, 0, 0, 0, 0, 0,
 	178, 173, 158, 134, 147, 132, 165, 187,
 	94, 100, 85, 67, 56, 53, 82, 84,
@@ -29,7 +29,7 @@ var egPawnTable = [64]int16{
 	13, 8, 8, 10, 13, 0, 2, -7,
 	0, 0, 0, 0, 0, 0, 0, 0}
 
-var mgKnightTable = [64]int16{
+var mgKnightTable = [64]int{
 	-167, -89, -34, -49, 61, -97, -15, -107,
 	-73, -41, 72, 36, 23, 62, 7, -17,
 	-47, 60, 37, 65, 84, 129, 73, 44,
@@ -39,7 +39,7 @@ var mgKnightTable = [64]int16{
 	-29, -53, -12, -3, -1, 18, -14, -19,
 	-105, -21, -58, -33, -17, -28, -19, -23}
 
-var egKnightTable = [64]int16{
+var egKnightTable = [64]int{
 	-58, -38, -13, -28, -31, -27, -63, -99,
 	-25, -8, -25, -2, -9, -25, -24, -52,
 	-24, -20, 10, 9, -1, -9, -19, -41,
@@ -49,7 +49,7 @@ var egKnightTable = [64]int16{
 	-42, -20, -10, -5, -2, -20, -23, -44,
 	-29, -51, -23, -15, -22, -18, -50, -64}
 
-var mgBishopTable = [64]int16{
+var mgBishopTable = [64]int{
 	-29, 4, -82, -37, -25, -42, 7, -8,
 	-26, 16, -18, -13, 30, 59, 18, -47,
 	-16, 37, 43, 40, 35, 50, 37, -2,
@@ -59,7 +59,7 @@ var mgBishopTable = [64]int16{
 	4, 15, 16, 0, 7, 21, 33, 1,
 	-33, -3, -14, -21, -13, -12, -39, -21}
 
-var egBishopTable = [64]int16{
+var egBishopTable = [64]int{
 	-14, -21, -11, -8, -7, -9, -17, -24,
 	-8, -4, 7, -12, -3, -13, -4, -14,
 	2, -8, 0, -1, -2, 6, 0, 4,
@@ -69,7 +69,7 @@ var egBishopTable = [64]int16{
 	-14, -18, -7, -1, 4, -9, -15, -27,
 	-23, -9, -23, -5, -9, -16, -5, -17}
 
-var mgRookTable = [64]int16{
+var mgRookTable = [64]int{
 	32, 42, 32, 51, 63, 9, 31, 43,
 	27, 32, 58, 62, 80, 67, 26, 44,
 	-5, 19, 26, 36, 17, 45, 61, 16,
@@ -79,7 +79,7 @@ var mgRookTable = [64]int16{
 	-44, -16, -20, -9, -1, 11, -6, -71,
 	-19, -13, 1, 17, 16, 7, -37, -26}
 
-var egRookTable = [64]int16{
+var egRookTable = [64]int{
 	13, 10, 18, 15, 12, 12, 8, 5,
 	11, 13, 13, 11, -3, 3, 8, 3,
 	7, 7, 7, 5, 4, -3, -5, -3,
@@ -89,7 +89,7 @@ var egRookTable = [64]int16{
 	-6, -6, 0, 2, -9, -9, -11, -3,
 	-9, 2, 3, -1, -5, -13, 4, -20}
 
-var mgQueenTable = [64]int16{
+var mgQueenTable = [64]int{
 	-28, 0, 29, 12, 59, 44, 43, 45,
 	-24, -39, -5, 1, -16, 57, 28, 54,
 	-13, -17, 7, 8, 29, 56, 47, 57,
@@ -99,7 +99,7 @@ var mgQueenTable = [64]int16{
 	-35, -8, 11, 2, 8, 15, -3, 1,
 	-1, -18, -9, 10, -15, -25, -31, -50}
 
-var egQueenTable = [64]int16{
+var egQueenTable = [64]int{
 	-9, 22, 22, 27, 27, 19, 10, 20,
 	-17, 20, 32, 41, 58, 25, 30, 0,
 	-20, 6, 9, 49, 47, 35, 19, 9,
@@ -109,7 +109,7 @@ var egQueenTable = [64]int16{
 	-22, -23, -30, -16, -16, -23, -36, -32,
 	-33, -28, -22, -43, -5, -32, -20, -41}
 
-var mgKingTable = [64]int16{
+var mgKingTable = [64]int{
 	-65, 23, 16, -15, -56, -34, 2, 13,
 	29, -1, -20, -7, -8, -4, -38, -29,
 	-9, 24, 2, -16, -20, 6, 22, -22,
@@ -119,7 +119,7 @@ var mgKingTable = [64]int16{
 	1, 7, -8, -64, -43, -16, 9, 8,
 	-15, 36, 12, -54, 8, -28, 24, 14}
 
-var egKingTable = [64]int16{
+var egKingTable = [64]int{
 	-74, -35, -18, -18, -11, 15, 4, -17,
 	-12, 17, 14, 17, 17, 38, 23, 11,
 	10, 17, 23, 15, 20, 45, 44, 13,
@@ -129,11 +129,11 @@ var egKingTable = [64]int16{
 	-27, -11, 4, 13, 14, 4, -5, -17,
 	-53, -34, -21, -11, -28, -14, -24, -43}
 
-var mgReferenceTable = [6]*[64]int16{&mgPawnTable, &mgKnightTable, &mgBishopTable, &mgRookTable, &mgQueenTable, &mgKingTable}
-var egReferenceTable = [6]*[64]int16{&egPawnTable, &egKnightTable, &egBishopTable, &egRookTable, &egQueenTable, &egKingTable}
+var mgReferenceTable = [6]*[64]int{&mgPawnTable, &mgKnightTable, &mgBishopTable, &mgRookTable, &mgQueenTable, &mgKingTable}
+var egReferenceTable = [6]*[64]int{&egPawnTable, &egKnightTable, &egBishopTable, &egRookTable, &egQueenTable, &egKingTable}
 
-var mgTable = [12][64]int16{}
-var egTable = [12][64]int16{}
+var mgTable = [12][64]int{}
+var egTable = [12][64]int{}
 
 func InitPeSTO() {
 	for p := PAWN; p <= KING; p++ {
@@ -146,7 +146,7 @@ func InitPeSTO() {
 	}
 }
 
-func (board *Board) Evaluate() (retval int16) {
+func (board *Board) Evaluate() (retval int) {
 	/******************
 	1. Piece evaluation
 	*******************/
@@ -158,9 +158,9 @@ func (board *Board) Evaluate() (retval int16) {
 }
 
 // Inspired heavily by https://www.chessprogramming.org/PeSTO%27s_Evaluation_Function, calculate with White being positive
-func PeSTOTableEval(board *Board) (retval, mgPhase, egPhase int16) {
-	var mg, eg [2]int16
-	var gamePhase int16
+func PeSTOTableEval(board *Board) (retval, mgPhase, egPhase int) {
+	var mg, eg [2]int
+	var gamePhase int
 
 	/* evaluate each piece */
 	for pos, piece := range board.PieceInfoArr {
